requirements.txt
PyQt5>=5.15
PyQt5-Qt5>=5.14
PyQt5_sip>=12.14
fastremap~=1.14.0
numpy~=1.26.4
numba==0.61.2
pysmb~= 1.2.9.1
joblib~=1.4.2
tqdm>=4.66.4
tifffile==2025.6.11
ipykernel~=6.29.4
scikit-image~=0.23.2
fpdf==1.7.2
seaborn==0.13.2
PyYAML>=6.0.1
matplotlib-scalebar==0.8.1
pydantic==2.7.4
joypy~=0.2.6
big-fish==0.6.2
trackpy>=0.6.3
readlif>=0.6.5
ipywidgets>=7.6.5
dna_features_viewer>=3.1.3
ipympl>=0.9.4
pystackreg>=0.2.7
snapgene-reader>=0.1.21
brokenaxes>=0.6.2
statsmodels==0.14.4
torch~=2.4.0
torchvision~=0.19.0
cellpose>=3.1.0
openpyxl>=3.1.5 
bioio>=1.6.1
bioio-lif~=1.2.0